[224] [220] Add user connection on a Services folder
